refactor: fix DEFAULT_SETTINGS circle reference
