+++ packages/const/package.json
@@ -9,13 +9,11 @@
   "main": "./src/index.ts",
   "dependencies": {
     "@icons-pack/react-simple-icons": "^13.8.0",
-    "@lobechat/business-config": "workspace:*",
     "@lobechat/business-const": "workspace:*",
     "@lobechat/types": "workspace:*",
     "@lobehub/ui": "^2.13.8",
     "klavis": "^2.15.0",
     "lodash-es": "^4.17.21",
-    "model-bank": "workspace:*",
     "query-string": "^9.3.1",
     "url-join": "^5.0.0"
   }
