+++ src/store/user/slices/settings/action.test.ts
@@ -1,9 +1,10 @@
+import { DEFAULT_SETTINGS } from '@lobechat/config';
+import { DEFAULT_AGENT } from '@lobechat/const';
 import { act, renderHook, waitFor } from '@testing-library/react';
 import type { PartialDeep } from 'type-fest';
 import { describe, expect, it, vi } from 'vitest';
 import { withSWR } from '~test-utils';
 
-import { DEFAULT_AGENT, DEFAULT_SETTINGS } from '@/const/settings';
 import { userService } from '@/services/user';
 import { useUserStore } from '@/store/user';
 import { LobeAgentSettings } from '@/types/session';
